requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.1
Pillow==10.0.1
psycopg2-binary==2.9.7
PyJWT==2.9.0
pyotp==2.9.0
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
python-decouple==3.8
whitenoise==6.6.0
django-cloudinary-storage==0.3.0
dj-database-url==2.1.0
cloudinary==1.39.0